--- a/analysis_options.yaml
+++ b/analysis_options.yaml
@@ -11,3 +11,4 @@ analyzer:
     - example/**
     - generator/**
     - benchmark/**
+    - lib/flatbuffers/**