--- a/benchmark/bin/main.dart
+++ b/benchmark/bin/main.dart
@@ -54,4 +54,3 @@ void main(List<String> arguments) {
     'removeAll',
   ]);
 }
-