… lib/flatbuffers/README.md
… lib/src/bindings/objectbox-c.dart
… lib/src/bindings/structs.dart
… lib/src/box.dart
… lib/src/modelinfo/entity_definition.dart
… lib/src/query/builder.dart
… lib/src/query/query.dart
… lib/src/store.dart
… pubspec.yaml
… test/box_test.dart
… test/flatbuffers_test.dart
… test/observer_test.dart
… test/query_property_test.dart