Tries to improve performance as evaluated by #144. Based on PR #163 - should be merged first. Also closes #62 

* [x] using [custom flatbuffers builder](https://github.com/objectbox/objectbox-dart/pull/164/commits/cc876475a195451cc9b5d44f604d01280c4653fe) with c-backed storage: putMany 110 ms -> 50 ms
* [x] [reusing flatbuffers builder & switching to cursor](https://github.com/objectbox/objectbox-dart/pull/164/commits/156457944f7789837b8c17324ee7e3b9139066f8) further improved putMany 50 ms -> 45 ms
* [x] [use FFI `memset()` to clear flatbuffers buffer on init/reset](https://github.com/objectbox/objectbox-dart/pull/164/commits/a96f6732b94ebfd91146a04a0dc03e91495209a2) instead of dart-only alternatives - putMany 45ms -> 40 ms
* [x] [group together value collection and flatbuffers marshalling](https://github.com/objectbox/objectbox-dart/pull/164/commits/534523cf7e1da333ff54526088df45a502093807) - putMany 40 ms -> 35 ms
* [x] use data visitors & generated unmarshalling code - readAll() 50 ms -> 10 ms