@vaind 2021-01-13
@greenrobot [back in 2019 you wanted to have a look before a merge](https://github.com/objectbox/objectbox-dart/issues/68#issuecomment-548369843)